全自动B站用户名组成歌词和视频渲染生成/requirements.txt
jieba==0.42.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
requests==2.26.0